+++ src/tankoh2/control/control_cl.py
@@ -17,7 +17,7 @@ from tankoh2.design.winding.windingutils import updateName, \
     changeSimulationOptions
 from tankoh2.design.winding.contour import getLiner, getDome  #, getLengthContourPath
 from tankoh2.geometry.geoutils import getReducedDomePoints
-from tankoh2.design.winding.material import getMaterial, getComposite, readLayupData
+from tankoh2.design.winding.material import getMaterial, getCompositeByLists, readLayupData
 from tankoh2.design.winding.optimize import optimizeFrictionGlobal_differential_evolution, optimizeHoopShiftForPolarOpeningX,\
     optimizeNegativeFrictionGlobal_differential_evolution
 from tankoh2.control.control_winding import createDesign
@@ -27,7 +27,7 @@ import tankoh2.design.existingdesigns
 
 def builtVesselAsBuilt(symmetricTank, servicepressure, saftyFactor, layersToWind, optimizeWindingHelical,
                        optimizeWindingHoop, tankname, dataDir, dcyl, polarOpening, lcylinder, dpoints,
-                       defaultLayerthickness, hoopLayerThickness, helixLayerThickenss, rovingWidth,
+                       defaultLayerthickness, layerThkHoop, layerThkHelical, rovingWidth,
                        numberOfRovingsHelical, numberOfRovingsHoop, tex, rho, hoopStart, hoopRisePerBandwidth,
                        minThicknessValue, hoopLayerCompressionStart, domeContourFilename):
     # #########################################################################################
@@ -88,13 +88,12 @@ def builtVesselAsBuilt(symmetricTank, servicepressure, saftyFactor, layersToWind
     log.info(f'get material')
     material = getMaterial(materialFilename)
 
-    
-
     angles, thicknesses, wendekreisradien, krempenradien, hoopShifts = readLayupData(layupDataFilename)
+    numberOfRovings = [numberOfRovingsHoop if angle > 88 else numberOfRovingsHelical for angle in angles]
     log.info(f'{angles[0:layersToWind]}')
-    composite = getComposite(angles[0:layersToWind], thicknesses[0:layersToWind], material, sectionAreaFibre,
-                             rovingWidth, numberOfRovingsHelical, numberOfRovingsHoop,
-                             tex, designFilename, tankname)
+    composite = getCompositeByLists(angles[0:layersToWind], thicknesses[0:layersToWind],
+                                    [rovingWidth] * layersToWind, numberOfRovings,
+                                    material, sectionAreaFibre, tex, designFilename, tankname)
 
     # create vessel and set liner and composite
     vessel = pychain.winding.Vessel()
@@ -347,8 +346,8 @@ def main():
     lcylinder = 500.  # mm    
     dpoints = 4  # data points for liner contour
     defaultLayerthickness = 0.17921146953405018 
-    hoopLayerThickness = 0.17921146953405018 
-    helixLayerThickenss = 0.17921146953405018  
+    layerThkHoop = 0.17921146953405018 
+    layerThkHelical = 0.17921146953405018  
     
     
     rovingWidth = 8
@@ -397,7 +396,7 @@ def main():
     if AsBuilt: 
         builtVesselAsBuilt(symmetricTank, servicepressure, safetyFactor, layersToWind, optimizeWindingHelical,
                            optimizeWindingHoop, tankname, dataDir, dcyl, polarOpening, lcylinder, dpoints,
-                           defaultLayerthickness, hoopLayerThickness, helixLayerThickenss, rovingWidth,
+                           defaultLayerthickness, layerThkHoop, layerThkHelical, rovingWidth,
                            numberOfRovingsHelical, numberOfRovingsHoop, tex, rho, hoopStart,
                            hoopRisePerBandwidth, minThicknessValue, hoopLayerCompressionStart, domeContourFilename)