+++ src/tankoh2/settings.py
@@ -8,6 +8,8 @@ from tankoh2.service.exception import Tankoh2Error
 myCrOSettings = None
 exampleSettingsFileName = 'settings_example.json'
 useRstOutput = False
+minCylindricalLength = 150
+epsilon = 1e-8
 
 optimizerSeed = None
 """Seed for evolutionary optimization for repeatable runs. If -1, no seed is used"""